Implemented inplace renaming for macros
